DIT-594 moving test to test folder
