+++ test/unit/views/AcceptItemInformationListViewSpec.scala
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2019 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package views
+
+import views.behaviours.ViewBehaviours
+import views.html.acceptItemInformationList
+
+class AcceptItemInformationListViewSpec extends ViewBehaviours {
+
+  val messageKeyPrefix = "acceptItemInformationList"
+
+  def createView = () => acceptItemInformationList(frontendAppConfig)(fakeRequest, messages)
+
+  "AcceptItemInformationList view" must {
+    behave like normalPage(createView, messageKeyPrefix)
+
+    behave like pageWithBackLink(createView)
+  }
+}
