+++ test/unit/views/behaviours/YesNoViewBehaviours.scala
@@ -31,7 +31,7 @@ trait YesNoViewBehaviours extends QuestionViewBehaviours[Boolean] {
           val doc = asDocument(createView(form))
           val legends = doc.getElementsByTag("legend")
           legends.size mustBe 1
-          legends.first.text mustBe messages(s"$messageKeyPrefix.heading")
+          legends.first.text mustBe messages(s"$messageKeyPrefix.label")
         }
 
         "contain an input for the value" in {
